dry getEnvironment code
